--- apps/webapp/src/types/vite-env.d.ts
@@ -1,1 +1,0 @@
-declare const __COMMIT_HASH__: string;
