+++ apps/webapp/src/vite-env.d.ts
@@ -1,3 +1,4 @@
 /// <reference types="vite/client" />
 declare const __COMMIT_HASH__: string;
 declare const __COMMIT_DATE__: string;
+declare const __GIT_ORIGIN_URL__: string;
